cfad/requirements.txt
torch>=2.0.0
torchmetrics>=1.0.0
Pillow>=9.5.0
opencv-python>=4.8.0
 PyYAML>=6.0
tqdm>=4.65.0
trimesh>=4.0.0
matplotlib>=3.7.0
numpy>=1.24.0
pillow>=9.5.0
